Switch from mcp-rb to fast-mcp for better protocol support
This major update transitions the Rails MCP Server from the mcp-rb
library to fast-mcp, enabling several key improvements:

- Added HTTP Server-Sent Events (SSE) support alongside STDIO mode
- Added new analysis tools:
  - analyze_controller_views: Examines relationships between controllers
and views
  - analyze_environment_config: Identifies security issues and
inconsistencies
- Improved command-line interface with mode selection (--mode flag)
- Added port configuration option (-p, --port)
- Refactored code into a more maintainable structure
- Updated documentation with detailed tool descriptions and examples

The server now provides two operation modes: STDIO (default) for direct
integration with clients like Claude Desktop, and HTTP mode with
JSON-RPC and SSE endpoints.

Bumps version to 1.1.0
