+++ lib/rails-mcp-server/version.rb
@@ -1,3 +1,3 @@
 module RailsMcpServer
-  VERSION = "1.0.1"
+  VERSION = "1.1.0"
 end
